pyproject.toml
[project]
name = "archipelagodiscordbot"
version = "0.1.0"
description = "Discord bot that logs Archipelago multiworld unlocks to a channel"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.7.1",
    "websockets>=13.0",
    "aiosqlite>=0.20",
    "cryptography>=43.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = ["-p", "no:cacheprovider"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/archibot"]
